+++ test/lib/types/application/ApplicationBuilder.js
@@ -67,7 +67,9 @@ test("Instantiation", (t) => {
 		"generateBundle",
 		"createDebugFiles",
 		"uglify",
-		"generateVersionInfo"
+		"generateVersionInfo",
+		"generateCachebusterInfo",
+		"generateApiIndex"
 	], "ApplicationBuilder is instantiated with standard tasks");
 });
 
@@ -87,7 +89,9 @@ test("Instantiation without component preload project configuration", (t) => {
 		"generateBundle",
 		"createDebugFiles",
 		"uglify",
-		"generateVersionInfo"
+		"generateVersionInfo",
+		"generateCachebusterInfo",
+		"generateApiIndex"
 	], "ApplicationBuilder is still instantiated with standard tasks");
 });
 
@@ -101,13 +105,13 @@ test("Instantiation without project namespace", (t) => {
 		"replaceCopyright",
 		"replaceVersion",
 		"generateFlexChangesBundle",
-		"generateManifestBundle",
 		"generateStandaloneAppBundle",
 		"transformBootstrapHtml",
 		"generateBundle",
 		"createDebugFiles",
 		"uglify",
-		"generateVersionInfo"
+		"generateVersionInfo",
+		"generateApiIndex"
 	], "All standard tasks but generateComponentPreload will be executed");
 });
 
@@ -132,6 +136,8 @@ test("Instantiation with custom tasks", (t) => {
 		"createDebugFiles",
 		"uglify",
 		"replaceVersion--1",
-		"generateVersionInfo"
+		"generateVersionInfo",
+		"generateCachebusterInfo",
+		"generateApiIndex"
 	], "ApplicationBuilder is still instantiated with standard tasks");
 });